Move map draw tools into useMapDrawTools hook
